backend/requirements.txt
fastapi[standard]
beanie
motor
python-decouple
bcrypt
